second commit

+++ Assignment1/Models/Product.cs
@@ -9,7 +9,7 @@ namespace Assignment1.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal Quantity { get; set; }
-        public decimal Price { get; set; }
+        public int Quantity { get; set; }
+        public int Price { get; set; }
     }
 }
